+++ src/Game/FrenchCards.hs
@@ -1,3 +1,3 @@
 module Game.FrenchCards (module Cards) where
 
-import Game.FrenchCards.Card as Cards
+import Game.FrenchCards.Card as Cards hiding (Card, cardWord)