+++ changelog.md
@@ -1,7 +1,7 @@
 # Change Log
 
 ## 0.2.0 (2016-09-16)
-* Card, newCard, cardRank, cardSuit, toWord, fromWord
+* Card, newCard, cardRank, cardSuit, cardWord
 * CardSet type, addCard, contains, suits
 * Pretty printing
 * Enum, Bounded